+++ backend/app/core/utils/query/auth_utils.py
