+++ backend/app/core/utils/query/portfolio_utils.py
