+++ backend/app/core/utils/query/stock_utils.py
